Add files via upload
Practice enterprise publicity website. It emulates to be part of a non-profit organization or a capital organization
I started adding some organizational info like Mision and Vision. After that I added a list of services, now it´s not functional, but soon it will
be a dynamic navbar.
At last there is a Team section, where i show the members profile photo, position and on click it will display more information.
And that´s it for this commit